+++ tests/server_curl.rs
@@ -3,23 +3,34 @@ use tls_tester::{NamedGroup, ServerCertificates, TlsServerBuilder, TlsServerStre
 use std::{
     io::{Read as _, Write as _},
     net::TcpListener,
-    process::{Command, Output},
+    process::Command,
 };
+
+// openssl req -x509 -newkey ec:<(openssl ecparam -name prime256v1) -keyout key_prime256v1.pem -out cert_prime256v1.pem -sha256 -days 3650 -nodes -subj "/CN=localhost"
+fn certs_ecdsa_secp256r1_sha256() -> ServerCertificates {
+    ServerCertificates::from_secpr256r1_pem("cert_prime256v1.pem", "key_prime256v1.pem")
+        .expect("Invalid certificates")
+}
+
+// openssl req -x509 -newkey rsa:2048 -nodes -keyout key_rsa_pss_rsae_sha256.pem -out cert_rsa_pss_rsae_sha256.pem -sha256 -days 3650 -subj "/CN=localhost"
+fn certs_rsa_pss_rsae_sha256() -> ServerCertificates {
+    ServerCertificates::from_rsa_pss_rsae_sha256(
+        "cert_rsa_pss_rsae_sha256.pem",
+        "key_rsa_pss_rsae_sha256.pem",
+    )
+    .expect("Invalid certificates")
+}
 
 fn test_curl_with_args(
     args: &'static [&'static str],
     supported_named_groups: Vec<NamedGroup>,
+    certs: ServerCertificates,
 ) -> TlsServerStream {
     stderrlog::new()
         .verbosity(4)
         .timestamp(stderrlog::Timestamp::Microsecond)
         .init()
         .ok();
-
-    // gen certs:
-    // openssl req -x509 -newkey ec:<(openssl ecparam -name prime256v1) -keyout key.pem -out cert.pem -sha256 -days 3650 -nodes -subj "/CN=localhost"
-    let certs: ServerCertificates = ServerCertificates::from_secpr256r1_pem("cert.pem", "key.pem")
-        .expect("Invalid certificates");
 
     let listener: TcpListener = TcpListener::bind("127.0.0.1:0").unwrap();
     let port: u16 = listener
@@ -29,7 +40,7 @@ fn test_curl_with_args(
 
     let curl_thread = std::thread::spawn(move || {
         log::info!("sending http request with curl");
-        let output: Output = Command::new("curl")
+        let mut output: std::process::Child = Command::new("curl")
             .arg(format!("https://127.0.0.1:{port}"))
             .arg("--tlsv1.3")
             .args(args)
@@ -40,23 +51,17 @@ fn test_curl_with_args(
             .arg("5")
             .arg("-v")
             .arg("--http1.1")
-            .output()
+            .spawn()
             .unwrap();
 
-        eprintln!("{}", String::from_utf8_lossy(&output.stderr));
-        let stdout = String::from_utf8_lossy(&output.stdout);
-        if !stdout.is_empty() {
-            panic!("{}", String::from_utf8_lossy(&output.stdout));
-        }
-
-        output.status.success()
+        output.wait().unwrap().success()
     });
 
     log::info!("Accepting connections");
     let (client, addr) = listener.accept().expect("Unable to accept connections");
     log::info!("Accepted connection from {addr}");
     let mut tls_stream: TlsServerStream = TlsServerBuilder::new()
-        .set_supported_name_groups(supported_named_groups)
+        .set_supported_named_groups(supported_named_groups)
         .handshake(client, certs)
         .expect("Failed to create TLS stream");
     let mut http_buf: [u8; 4096] = [0; 4096];
@@ -79,12 +84,20 @@ fn test_curl_with_args(
 
 #[test]
 fn server_curl_secp256r1() {
-    test_curl_with_args(&["--curves", "secp256r1"], vec![NamedGroup::secp256r1]);
+    test_curl_with_args(
+        &["--curves", "secp256r1"],
+        vec![NamedGroup::secp256r1],
+        certs_ecdsa_secp256r1_sha256(),
+    );
 }
 
 #[test]
 fn server_curl_x25519() {
-    test_curl_with_args(&["--curves", "x25519"], vec![NamedGroup::x25519]);
+    test_curl_with_args(
+        &["--curves", "x25519"],
+        vec![NamedGroup::x25519],
+        certs_ecdsa_secp256r1_sha256(),
+    );
 }
 
 #[test]
@@ -92,10 +105,25 @@ fn server_curl_hello_retry() {
     let server_stream = test_curl_with_args(
         &["--curves", "x25519:secp256r1"],
         vec![NamedGroup::secp256r1],
+        certs_ecdsa_secp256r1_sha256(),
     );
 
     assert!(
         server_stream.hello_retry_request(),
         "Server did not send a hello retry"
+    );
+}
+
+#[test]
+fn server_curl_rsa_pss_rsae_sha256() {
+    stderrlog::new()
+        .verbosity(4)
+        .timestamp(stderrlog::Timestamp::Microsecond)
+        .init()
+        .ok();
+    test_curl_with_args(
+        &["--curves", "x25519", "--sigalgs", "rsa_pss_rsae_sha256"],
+        vec![NamedGroup::x25519],
+        certs_rsa_pss_rsae_sha256(),
     );
 }